feat(auth): seperate auth api from user
