+++ DI/containers.py
@@ -2,6 +2,7 @@ from dependency_injector import containers, providers
 
 from config import get_settings
 from db.database import AsyncSessionLocal
+from user.application.auth_service import AuthService
 from user.application.user_service import UserService
 from user.infra.repository.user_repo import UserRepository
 from ulid import ULID
@@ -12,7 +13,9 @@ from utils.crypto import Crypto
 class Container(containers.DeclarativeContainer):
     # setting
     wiring_config = containers.WiringConfiguration(
-        packages=["user.interface.controllers"]
+        packages=[
+            "user.interface.controllers",
+        ]
     )
 
     config = providers.Singleton(get_settings)
@@ -26,6 +29,13 @@ class Container(containers.DeclarativeContainer):
 
     user_service = providers.Factory(
         UserService,
+        user_repo=user_repository,
+        ulid=providers.Singleton(ULID),
+        crypto=providers.Singleton(Crypto),
+    )
+
+    auth_service = providers.Factory(
+        AuthService,
         user_repo=user_repository,
         ulid=providers.Singleton(ULID),
         crypto=providers.Singleton(Crypto),
